+++ src/main/java/com/kakaotech/team14backend/point/domain/PointHistory.java
@@ -1,15 +1,16 @@
-package com.kakaotech.team14backend.inner.point.model;
+package com.kakaotech.team14backend.point.domain;
 
-import static lombok.AccessLevel.PROTECTED;
-
-import java.time.LocalDateTime;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import java.time.LocalDateTime;
+
+import static lombok.AccessLevel.PROTECTED;
 
 @Entity(name = "point_history")
 @NoArgsConstructor(access = PROTECTED)
