+++ src/main/java/com/kakaotech/team14backend/point/domain/UsePointDecider.java
@@ -1,4 +1,4 @@
-package com.kakaotech.team14backend.inner.point.model;
+package com.kakaotech.team14backend.point.domain;
 
 import com.kakaotech.team14backend.post.exception.PostLevelOutOfRangeException;
